0.0.14 combinators.txt touch-ups, pondering refactoring
Although treating obAB extensionally as simply an Ob subset (except for the implied order) is accomplished, I am concerned about the order of the narrative with regard to avoiding non-deterministic or lazy operation and where deterministic elimination of unnecessary computations is addressed.  This sets the stage for resolution of that.